Design: PerfumeView 분리 및 디자인 수정
